# Remove hash hacks