We can use a `runtime.Pinner` and the `noescape`/`nocallback` directives to replace some hacky logic that we implemented in the several hash functions to avoid allocations and to workaround a `cgoCheckPointer` false positive (tested in `TestCgo`).

Performance is not affected:

```
goos: linux
goarch: amd64
pkg: github.com/golang-fips/openssl/v2
cpu: Intel(R) Core(TM) i7-10850H CPU @ 2.70GHz
                │   old.txt    │              new.txt               │
                │    sec/op    │   sec/op     vs base               │
SHA256-4          814.0n ±  3%   792.2n ± 6%       ~ (p=0.105 n=10)
SHA256OneShot-4   585.8n ± 15%   607.1n ± 4%       ~ (p=0.089 n=10)
geomean           690.5n         693.5n       +0.42%

                │    old.txt    │               new.txt               │
                │      B/s      │     B/s       vs base               │
SHA256-4          9.370Mi ±  2%   9.632Mi ± 6%       ~ (p=0.093 n=10)
SHA256OneShot-4   13.02Mi ± 18%   12.56Mi ± 4%       ~ (p=0.089 n=10)
geomean           11.05Mi         11.00Mi       -0.41%

                │   old.txt    │               new.txt               │
                │     B/op     │    B/op     vs base                 │
SHA256-4          64.00 ± 0%     64.00 ± 0%       ~ (p=1.000 n=10) ¹
SHA256OneShot-4   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                      ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                │   old.txt    │               new.txt               │
                │  allocs/op   │ allocs/op   vs base                 │
SHA256-4          1.000 ± 0%     1.000 ± 0%       ~ (p=1.000 n=10) ¹
SHA256OneShot-4   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                      ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```